Add renumber_faces() to renumber the indices to vertices in faces.
The program renumbers the faces lists' indices to vertices,
so that the new indices are in the range of the number of indices:
range(len(indices)). This assumes that all of the indices are
represented in the faces, for example, after running remove_faces().

This function was introduced to prepare faces for fem_laplacian().


Former-commit-id: b26356f059ea8d0333975159b231ef9e63668c3c